cardano-private-network/docker/didcomm-server-docker/didcomm-server/requirements.txt
aiohttp==3.8.3
aiohttp-apispec==2.2.3
aiohttp-cors==0.7.0
aiosignal==1.2.0
base58==2.1.1
bcrypt==4.0.1
cryptography==36.0.0
didcomm==0.3.0
ecdsa==0.18.0
fastapi==0.85.0
graphene==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
httptools==0.5.0
jwt==1.3.1
numpy==1.22.0
peerdid==0.4.0
PyJWT==2.5.0
pymongo==4.2.0
python-multipart==0.0.5
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
strawberry-graphql==0.133.1
typing-extensions==3.10.0.2
uvicorn==0.18.3
